tappy/apps/api/pyproject.toml
[project]
name = "tappy-api"
version = "0.1.0"
description = "FastAPI service for Tappy journal automation"
authors = [{name = "Tappy"}]
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.5",
  "google-genai>=0.1.0",
  "httpx>=0.27.0",
  "pydantic>=2.9.2",
  "uvicorn[standard]>=0.32.1",
  "sqlalchemy[asyncio]>=2.0.0",
  "aiosqlite>=0.20.0",
  "python-dotenv>=1.0.0",
  "browser-use>=0.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
